kernel D works

+++ image/image.cpp
@@ -6,7 +6,7 @@ int image_init (Image* img, int width, int height)
 {
    img->width = width;
    img->height = height;
-   img->data = (byte*)malloc(width * height * sizeof(byte));
+   img->data = (int*)malloc(width * height * sizeof(int));
    if (!img->data)
       return 1;
    return 0;
@@ -34,7 +34,7 @@ int image_load (Image* img, const char* filename)
    fread(&img->ydpi, sizeof(int), 1, f);
    for (y = 0; y < img->height; ++y)
       for (x = 0; x < img->width; ++x)
-         fread(&img->data[y * img->width + x], sizeof(byte), 1, f);
+         fread(&img->data[y * img->width + x], sizeof(int), 1, f);
    return 0;
 }
 
@@ -45,7 +45,7 @@ int image_save (const char* filename, Image* img)
    fwrite(&img->height, sizeof(int), 1, f);
    fwrite(&img->xdpi, sizeof(int), 1, f);
    fwrite(&img->ydpi, sizeof(int), 1, f);
-   fwrite(img->data, sizeof(byte), img->width * img->height, f);
+   fwrite(img->data, sizeof(int), img->width * img->height, f);
    fclose(f);
    return 0;
 }
